Export a function instead of a data source
The data source implemented in previous versions of this provider always
really just wanted to be a function, but that wasn't possible in Terraform
v1.7 and earlier.

Terraform v1.8 allows providers to extend the Terraform language with new
functions, and so this is a total rework of this provider exporting a
function named "script" in place of the old data source. The functionality
is still essentially the same, but should now be more convenient to use.